+++ apps/dashboard/app/(app)/semantic-cache/[gatewayId]/logs/page.tsx
@@ -1,48 +1,13 @@
-// import { PageHeader } from "@/components/dashboard/page-header";
-import { Button } from "@/components/ui/button";
-// import {
-//   Dialog,
-//   DialogContent,
-//   DialogDescription,
-//   DialogFooter,
-//   DialogHeader,
-//   DialogTitle,
-//   DialogTrigger,
-// } from "@/components/ui/dialog";
-// import { Input } from "@/components/ui/input";
-// import { Label } from "@/components/ui/label";
-// import { Separator } from "@/components/ui/separator";
-
-// import {
-//   Table,
-//   TableBody,
-//   TableCaption,
-//   TableCell,
-//   TableHead,
-//   TableHeader,
-//   TableRow,
-// } from "@/components/ui/table";
-
 import { getTenantId } from "@/lib/auth";
 import { db } from "@/lib/db";
 import { getAllSemanticCacheLogs } from "@/lib/tinybird";
 import { redirect } from "next/navigation";
-import { IntervalSelect } from "../../../apis/[apiId]/select";
-import Table from "./table";
+import { LogsTable } from "./table";
+import { getInterval } from "./util";
 
-// const formatDate = (timestamp: string | number | Date): string => {
-//   const date = new Date(timestamp);
-//   const options: Intl.DateTimeFormatOptions = {
-//     month: "long",
-//     day: "numeric",
-//     hour: "2-digit",
-//     minute: "2-digit",
-//     second: "2-digit",
-//   };
-//   return date.toLocaleDateString("en-US", options);
-// };
-
-export default async function SemanticCacheLogsPage() {
+export default async function SemanticCacheLogsPage({
+  searchParams,
+}: { searchParams: { interval?: string } }) {
   const tenantId = getTenantId();
   const workspace = await db.query.workspaces.findFirst({
     where: (table, { and, eq, isNull }) =>
@@ -67,11 +32,14 @@ export default async function SemanticCacheLogsPage() {
     return redirect("/semantic-cache/new");
   }
 
+  const interval = getInterval(searchParams.interval || "7d");
+
   const { data } = await getAllSemanticCacheLogs({
     gatewayId,
-    workspaceId: workspace.id,
+    workspaceId: workspace?.id,
     limit: 1000,
+    interval,
   });
 
-  return <Table data={data} workspace={workspace} />;
+  return <LogsTable data={data} workspace={workspace} />;
 }
